+++ src/daft/exceptions.py
